backend-legacy/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pypdf
requests
beautifulsoup4
rank-bm25
openai